[Refactor] Refactor memberUpdate page
